Tests/LibWeb/Text/expected/abspos-relayout-insert-table-row.txt

live-new=0,20,50,20
outside-before=0,20,50,20
outside-after=0,20,50,20
partial-layouts=0 full-layouts=1
FAIL
